pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prd-pal"
version = "0.6.0"
description = "prd-pal review engine for requirement analysis, handoff preparation, and delivery support"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
  { name = "Project Maintainers" }
]
dependencies = [
  "aiofiles>=23.2.1",
  "colorama>=0.4.6",
  "fastapi>=0.104.1",
  "httpx>=0.27.0",
  "json-repair>=0.29.8",
  "langchain-community>=0.4.0",
  "langchain-core>=1.0.0",
  "langchain-openai>=1.0.0",
  "langchain-ollama>=1.0.0",
  "langgraph>=0.2.76",
  "litellm>=1.71.0",
  "mcp>=1.9.1",
  "pydantic>=2.5.1",
  "python-dotenv>=1.0.0",
  "python-multipart>=0.0.6",
  "tiktoken>=0.7.0",
  "uvicorn>=0.24.0.post1"
]

[project.optional-dependencies]
test = [
  "pytest-asyncio>=0.23.0"
]

[project.scripts]
prd-pal = "prd_pal.main:run_cli"
prd-review = "prd_pal.main:run_cli"

[tool.setuptools.packages.find]
include = ["prd_pal*", "requirement_review_v1*", "review_runtime*"]

[tool.setuptools]
license-files = ["LICENSE", "NOTICE"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
addopts = "-v"
testpaths = ["tests"]
python_files = "test_*.py"
